backend/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-crispy-forms==1.11.2
django-filter==2.4.0
djangorestframework==3.12.4
importlib-metadata==4.3.1
Markdown==3.3.4
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
zipp==3.4.1
